Menu integration and copy to clipboard util
Done the integration between the menu and validations and also the same way between menu and search. 
Search for Name and Extesion is working for now, we will change the CLI - UI in order to add support to an advanced search ino rder to support the rest of  options.
Add a util to copy to clipboard any text